Fix ADP violation
