+++ jme3-core/src/main/resources/com/jme3/system/version.properties
@@ -1,12 +1,12 @@
 # THIS IS AN AUTO-GENERATED FILE..
 # DO NOT MODIFY!
-build.date=2016-03-24
-git.revision=5477
+build.date=2016-04-10
+git.revision=5483
 git.branch=master
-git.hash=16d3ca509e9113250a13138502e4e81f843ee796
-git.hash.short=16d3ca5
+git.hash=bbfdbbfdfe04fa46dd6536afc799f54a2e6db084
+git.hash.short=bbfdbbf
 git.tag=null
-name.full=jMonkeyEngine 3.1-5477
-version.full=3.1-5477
+name.full=jMonkeyEngine 3.1-5483
+version.full=3.1-5483
 version.number=3.1.0
 version.tag=SNAPSHOT
