+++ jme3-examples/src/main/java/jme3test/texture/TestImageRaster.java
@@ -19,6 +19,7 @@ import com.jme3.texture.Texture.MinFilter;
 import com.jme3.texture.Texture2D;
 import com.jme3.texture.image.ColorSpace;
 import com.jme3.texture.image.ImageRaster;
+import com.jme3.texture.image.ImageRasterFactory;
 import com.jme3.util.BufferUtils;
 import java.nio.ByteBuffer;
 
@@ -30,8 +31,8 @@ public class TestImageRaster extends SimpleApplication {
         ByteBuffer data = BufferUtils.createByteBuffer( (int)Math.ceil(newFormat.getBitsPerPixel() / 8.0) * width * height);
         Image convertedImage = new Image(newFormat, width, height, data,null, image.getColorSpace());
         
-        ImageRaster sourceReader = ImageRaster.create(image);
-        ImageRaster targetWriter = ImageRaster.create(convertedImage);
+        ImageRaster sourceReader = ImageRasterFactory.create(image);
+        ImageRaster targetWriter = ImageRasterFactory.create(convertedImage);
         for (int x = 0; x < width; x++) {
             for (int y = 0; y < height; y++) {
                 ColorRGBA color = sourceReader.getPixel(x, y);
@@ -69,7 +70,7 @@ public class TestImageRaster extends SimpleApplication {
     private Image createTestImage() {
         Image testImage = new Image(Format.BGR8, 4, 3, BufferUtils.createByteBuffer(4 * 4 * 3), null, ColorSpace.Linear);
         
-        ImageRaster io = ImageRaster.create(testImage);
+        ImageRaster io = ImageRasterFactory.create(testImage);
         io.setPixel(0, 0, ColorRGBA.Black);
         io.setPixel(1, 0, ColorRGBA.Gray);
         io.setPixel(2, 0, ColorRGBA.White);
